Fixed dewy utils
